Revert "Release 252.25 (#78)"
This reverts commit c094b19c75a5d11e2484a3ecbebe5fb579499c5e.
