+++ gradle.properties
@@ -1,7 +1,4 @@
 org.gradle.jvmargs=-Xmx2048M
-org.gradle.configuration-cache=true
-org.gradle.configuration-cache.problems=warn
-org.gradle.caching=true
 kotlin.stdlib.default.dependency=false
 
 kotlinVersion=2.1.20
