+++ modules/jpa/src/main/resources/messages/JpaBundle.properties
@@ -8,4 +8,3 @@ explyt.jpa.inspection.insert.parameters.count.lack=Lack of insertion parameters.
 explyt.jpa.inspection.insert.parameters.count.unexpected=Unexpected insertion parameter. The {0} is expected.
 explyt.jpa.inspection.insert.type.mismatch=Invalid expression type. {0} is expected, but {1} is received.
 explyt.jpa.usage.type=JPA query usage
-explyt.jpa.action.eq.hashcode.title=JPA: equals() and hashCode()
