--- modules/spring-ai/spring-ai.gradle.kts
@@ -1,76 +1,0 @@
-/*
- * Copyright © 2025 Explyt Ltd
- *
- * All rights reserved.
- *
- * This code and software are the property of Explyt Ltd and are protected by copyright and other intellectual property laws.
- *
- * You may use this code under the terms of the Explyt Source License Version 1.0 ("License"), if you accept its terms and conditions.
- *
- * By installing, downloading, accessing, using, or distributing this code, you agree to the terms and conditions of the License.
- * If you do not agree to such terms and conditions, you must cease using this code and immediately delete all copies of it.
- *
- * You may obtain a copy of the License at: https://github.com/explyt/spring-plugin/blob/main/EXPLYT-SOURCE-LICENSE.md
- *
- * Unauthorized use of this code constitutes a violation of intellectual property rights and may result in legal action.
- */
-
-import org.jetbrains.intellij.platform.gradle.TestFrameworkType
-
-plugins {
-    java
-    id("org.jetbrains.intellij.platform.module")
-    kotlin("jvm")
-}
-
-evaluationDependsOn(":spring-core")
-evaluationDependsOn(":spring-web")
-
-val intellijPlugins = listOf<String>()
-
-ext {
-    set("intellijPlugins", intellijPlugins)
-}
-
-val springCoreProject = project(":spring-core")
-val springWebProject = project(":spring-web")
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_21
-}
-
-intellijPlatform {
-    buildSearchableOptions = false
-    instrumentCode = false
-}
-
-
-tasks {
-    runIde { enabled = false }
-    publishPlugin { enabled = false }
-    verifyPlugin { enabled = false }
-}
-
-val defaultIdeaType: String by rootProject
-val defaultIdeaVersion: String by rootProject
-
-dependencies {
-    implementation(springCoreProject)
-    implementation(springWebProject)
-
-    intellijPlatform {
-        create(defaultIdeaType, defaultIdeaVersion, useInstaller = false)
-        jetbrainsRuntime()
-        bundledPlugins(springCoreProject.ext["intellijPlugins"] as List<String> + intellijPlugins)
-        testFramework(TestFrameworkType.Platform)
-        testFramework(TestFrameworkType.Plugin.Java)
-
-        plugin("com.explyt.test", "4.1.3-IJ-251")
-    }
-    testImplementation(project(":test-framework"))
-    testImplementation("junit:junit:4.13.2")
-}
-
-kotlin {
-    jvmToolchain(21)
-}
