--- modules/spring-ai/src/main/kotlin/com/explyt/spring/ai/provider/HttpAiActionPanelLoader.kt
@@ -1,35 +1,0 @@
-/*
- * Copyright © 2025 Explyt Ltd
- *
- * All rights reserved.
- *
- * This code and software are the property of Explyt Ltd and are protected by copyright and other intellectual property laws.
- *
- * You may use this code under the terms of the Explyt Source License Version 1.0 ("License"), if you accept its terms and conditions.
- *
- * By installing, downloading, accessing, using, or distributing this code, you agree to the terms and conditions of the License.
- * If you do not agree to such terms and conditions, you must cease using this code and immediately delete all copies of it.
- *
- * You may obtain a copy of the License at: https://github.com/explyt/spring-plugin/blob/main/EXPLYT-SOURCE-LICENSE.md
- *
- * Unauthorized use of this code constitutes a violation of intellectual property rights and may result in legal action.
- */
-
-package com.explyt.spring.ai.provider
-
-import com.explyt.spring.ai.action.ConvertCurlToHttpAction
-import com.explyt.spring.ai.action.ConvertPostmanToHttpAction
-import com.explyt.spring.web.httpclient.AdditionalHttpPanelLoader
-import com.intellij.openapi.ui.DialogPanel
-import com.intellij.ui.dsl.builder.panel
-
-class HttpAiActionPanelLoader : AdditionalHttpPanelLoader {
-    override fun getPanel(): DialogPanel {
-        return panel {
-            row {
-                button("Import from Postman", ConvertPostmanToHttpAction())
-                button("From Curl", ConvertCurlToHttpAction())
-            }
-        }
-    }
-}
