--- modules/spring-ai/src/main/resources/META-INF/spring-ai-plugin.xml
@@ -1,61 +1,0 @@
-<!--
-  ~ Copyright © 2025 Explyt Ltd
-  ~
-  ~ All rights reserved.
-  ~
-  ~ This code and software are the property of Explyt Ltd and are protected by copyright and other intellectual property laws.
-  ~
-  ~ You may use this code under the terms of the Explyt Source License Version 1.0 ("License"), if you accept its terms and conditions.
-  ~
-  ~ By installing, downloading, accessing, using, or distributing this code, you agree to the terms and conditions of the License.
-  ~ If you do not agree to such terms and conditions, you must cease using this code and immediately delete all copies of it.
-  ~
-  ~ You may obtain a copy of the License at: https://github.com/explyt/spring-plugin/blob/main/EXPLYT-SOURCE-LICENSE.md
-  ~
-  ~ Unauthorized use of this code constitutes a violation of intellectual property rights and may result in legal action.
-  -->
-
-<idea-plugin>
-    <depends optional="true" config-file="spring-test-plugin.xml">com.explyt.test</depends>
-
-    <extensions defaultExtensionNs="com.intellij">
-        <postStartupActivity implementation="com.explyt.spring.ai.service.AiStartupActivity"/>
-    </extensions>
-    <actions>
-        <group id="ExternalCallMenuGroup"
-               text="Explyt Spring AI Actions"
-               icon="/com/explyt/ai/icons/assistantResponseIcon.svg"
-               popup="true"
-               description="Refactor actions">
-            <add-to-group group-id="EditorPopupMenu" anchor="last"/>
-            <add-to-group group-id="ProjectViewPopupMenu" anchor="last"/>
-
-            <action id="Explyt.Spring.ConvertPropertiesToYamlAction"
-                    class="com.explyt.spring.ai.action.ConvertPropertiesToYamlAction"/>
-
-            <action id="Explyt.Spring.ConvertYamlToPropertiesAction"
-                    class="com.explyt.spring.ai.action.ConvertYamlToPropertiesAction"/>
-
-            <action id="Explyt.Spring.ConvertDtoToEntityAction"
-                    class="com.explyt.spring.ai.action.ConvertDtoToEntityAction"/>
-
-            <action id="Explyt.Spring.ConvertEntityToDtoAction"
-                    class="com.explyt.spring.ai.action.ConvertEntityToDtoAction"/>
-
-            <action id="Explyt.Spring.ConvertControllerToOpenapiAction"
-                    class="com.explyt.spring.ai.action.ConvertControllerToOpenapiAction"/>
-
-            <action id="Explyt.Spring.ConvertOpenapiToControllerAction"
-                    class="com.explyt.spring.ai.action.ConvertOpenapiToControllerAction"/>
-
-            <action id="Explyt.Spring.ConvertCurlToHttpAction"
-                    class="com.explyt.spring.ai.action.ConvertCurlToHttpAction"/>
-
-            <action id="Explyt.Spring.ConvertPostmanToHttpAction"
-                    class="com.explyt.spring.ai.action.ConvertPostmanToHttpAction"/>
-
-            <action id="Explyt.Spring.ConvertDbMigrationFilesToToEntityAction"
-                    class="com.explyt.spring.ai.action.ConvertDbMigrationFilesToEntityAction"/>
-        </group>
-    </actions>
-</idea-plugin>
