--- modules/spring-ai/src/main/resources/META-INF/spring-test-plugin.xml
@@ -1,36 +1,0 @@
-<!--
-  ~ Copyright © 2025 Explyt Ltd
-  ~
-  ~ All rights reserved.
-  ~
-  ~ This code and software are the property of Explyt Ltd and are protected by copyright and other intellectual property laws.
-  ~
-  ~ You may use this code under the terms of the Explyt Source License Version 1.0 ("License"), if you accept its terms and conditions.
-  ~
-  ~ By installing, downloading, accessing, using, or distributing this code, you agree to the terms and conditions of the License.
-  ~ If you do not agree to such terms and conditions, you must cease using this code and immediately delete all copies of it.
-  ~
-  ~ You may obtain a copy of the License at: https://github.com/explyt/spring-plugin/blob/main/EXPLYT-SOURCE-LICENSE.md
-  ~
-  ~ Unauthorized use of this code constitutes a violation of intellectual property rights and may result in legal action.
-  -->
-
-<idea-plugin>
-    <extensions defaultExtensionNs="com.explyt.spring.web">
-        <additionalHttpPanelLoader implementation="com.explyt.spring.ai.provider.HttpAiActionPanelLoader"/>
-    </extensions>
-    <extensions defaultExtensionNs="com.intellij">
-        <codeInsight.lineMarkerProvider
-                language="UAST"
-                implementationClass="com.explyt.spring.ai.provider.SpringAiActionsLineMarkerProvider"/>
-
-        <codeInsight.lineMarkerProvider
-                language="yaml"
-                implementationClass="com.explyt.spring.ai.provider.YamlOpenApiActionsLineMarkerProvider"/>
-        <codeInsight.lineMarkerProvider
-                language="JSON"
-                implementationClass="com.explyt.spring.ai.provider.JsonOpenApiActionsLineMarkerProvider"/>
-    </extensions>
-    <extensions defaultExtensionNs="org.jetbrains.kotlin">
-    </extensions>
-</idea-plugin>
