+++ modules/spring-bootstrap/spring-bootstrap.gradle.kts
@@ -38,7 +38,6 @@ val allCompileProjects = listOf(
     "spring-aop",
     "jpa",
     "quarkus-core",
-    "spring-ai",
 )
 
 allCompileProjects.forEach {
