+++ modules/spring-bootstrap/src/main/resources/META-INF/plugin.xml
@@ -52,5 +52,4 @@
     <xi:include href="jpa-plugin.xml" xpointer="xpointer(/idea-plugin/*)"/>
     <xi:include href="spring-gradle-plugin.xml" xpointer="xpointer(/idea-plugin/*)"/>
     <xi:include href="quarkus-core-plugin.xml" xpointer="xpointer(/idea-plugin/*)"/>
-    <xi:include href="spring-ai-plugin.xml" xpointer="xpointer(/idea-plugin/*)"/>
 </idea-plugin>
