+++ modules/spring-core/src/main/kotlin/com/explyt/spring/core/action/SpringJavaMethodGenerateAction.kt
@@ -24,21 +24,24 @@ import com.explyt.spring.core.statistic.StatisticActionId.GENERATE_POST_CONSTRUC
 import com.explyt.spring.core.statistic.StatisticActionId.GENERATE_PRE_DESTROY
 import com.explyt.spring.core.statistic.StatisticService
 import com.explyt.spring.core.util.SpringCoreUtil
-import com.explyt.util.JavaMethodGenerateUtils
 import com.intellij.codeInsight.CodeInsightActionHandler
+import com.intellij.codeInsight.daemon.impl.quickfix.CreateFromUsageUtils
 import com.intellij.codeInsight.generation.actions.BaseGenerateAction
 import com.intellij.codeInsight.template.Template
+import com.intellij.codeInsight.template.TemplateEditingAdapter
 import com.intellij.codeInsight.template.TemplateManager
 import com.intellij.codeInsight.template.impl.ConstantNode
+import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.runWriteAction
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.module.Module
 import com.intellij.openapi.module.ModuleUtil
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.text.StringUtil
-import com.intellij.psi.PsiClass
-import com.intellij.psi.PsiDocumentManager
-import com.intellij.psi.PsiFile
+import com.intellij.psi.*
+import com.intellij.psi.util.PsiTreeUtil
+import com.intellij.psi.util.startOffset
+import org.jetbrains.kotlin.psi.psiUtil.endOffset
 
 
 class PostConstructJavaGenerateAction : BaseGenerateAction(PostConstructMethodHandler()) {
@@ -70,7 +73,7 @@ private class PostConstructMethodHandler(val postConstruct: Boolean = true) : Co
         val actionId = if (postConstruct) GENERATE_POST_CONSTRUCT else GENERATE_PRE_DESTROY
         StatisticService.getInstance().addActionUsage(actionId)
         val module = ModuleUtil.findModuleForPsiElement(file) ?: return
-        val targetClass = JavaCoreMethodGenerateUtils.getNearTargetClass(editor, file) ?: return
+        val targetClass = JavaMethodGenerateUtils.getNearTargetClass(editor, file) ?: return
 
         runWriteAction {
             val documentManager = PsiDocumentManager.getInstance(project)
@@ -114,9 +117,58 @@ private class PostConstructMethodHandler(val postConstruct: Boolean = true) : Co
     }
 }
 
-object JavaCoreMethodGenerateUtils {
+object JavaMethodGenerateUtils {
 
     fun getNearTargetClass(editor: Editor?, file: PsiFile?): PsiClass? {
         return PostConstructJavaGenerateAction().getNearTargetClass(editor, file)
+    }
+
+    fun findOffsetToInsertMethod(editor: Editor, file: PsiFile, targetClass: PsiClass): Int {
+        var result = editor.caretModel.offset
+        val psiMethod = PsiTreeUtil.findElementOfClassAtOffset(file, result - 1, PsiMethod::class.java, false)
+        if (psiMethod != null) {
+            return psiMethod.startOffset
+        }
+        val psiField = PsiTreeUtil.findElementOfClassAtOffset(file, result - 1, PsiField::class.java, false)
+        if (psiField != null) {
+            return psiField.endOffset
+        }
+
+        var classAtCursor = PsiTreeUtil.getParentOfType(file.findElementAt(result), PsiClass::class.java, false)
+        if (classAtCursor == targetClass) {
+            return result
+        }
+
+        while (classAtCursor != null && classAtCursor.parent !is PsiFile) {
+            result = classAtCursor.textRange.endOffset
+            classAtCursor = PsiTreeUtil.getParentOfType(classAtCursor, PsiClass::class.java)
+        }
+
+        return result
+    }
+
+    fun startTemplate(project: Project, editor: Editor, template: Template) {
+        val adapter = object : TemplateEditingAdapter() {
+            override fun templateFinished(template: Template, brokenOff: Boolean) {
+                ApplicationManager.getApplication().runWriteAction {
+                    PsiDocumentManager.getInstance(project).commitDocument(editor.document)
+                    PsiDocumentManager.getInstance(project).getPsiFile(editor.document)
+                        ?.let { findPsiMethod(it, editor) }
+                        ?.let { PsiTreeUtil.getParentOfType(it, PsiMethod::class.java, false) }
+                        ?.let { CreateFromUsageUtils.setupEditor(it, editor) }
+                }
+            }
+        }
+        TemplateManager.getInstance(project).startTemplate(editor, template, adapter)
+    }
+
+    private fun findPsiMethod(it: PsiFile, editor: Editor): PsiMethod? {
+        for (i in 2..20) {
+            val psiMethod = PsiTreeUtil.findElementOfClassAtOffset(
+                it, editor.caretModel.offset - i, PsiMethod::class.java, false
+            )
+            if (psiMethod != null) return psiMethod
+        }
+        return null
     }
 }
