+++ modules/spring-core/src/main/kotlin/com/explyt/spring/core/action/UastModelTrackerInvalidateAction.kt
@@ -21,7 +21,6 @@ import com.explyt.spring.core.SpringCoreBundle
 import com.explyt.spring.core.SpringIcons
 import com.explyt.spring.core.service.SpringSearchServiceFacade
 import com.explyt.spring.core.tracker.ModificationTrackerManager
-import com.explyt.spring.core.util.ActionUtil
 import com.intellij.codeInsight.daemon.DaemonCodeAnalyzer
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
@@ -29,7 +28,6 @@ import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
 import com.intellij.openapi.module.ModuleUtilCore
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.project.guessProjectDir
 import com.intellij.psi.PsiManager
 
 class UastModelTrackerInvalidateAction : AnAction() {
@@ -48,26 +46,13 @@ class UastModelTrackerInvalidateAction : AnAction() {
         val project = e.project
         val p = e.presentation
         if (project == null) {
-            ActionUtil.isEnabledAndVisible(e, false)
-            return
-        }
-        val virtualFiles = e.getData(CommonDataKeys.VIRTUAL_FILE_ARRAY)
-        if (virtualFiles == null || virtualFiles.size > 1) {
-            ActionUtil.isEnabledAndVisible(e, false)
+            p.isVisible = false
             return
         }
         val file = e.getData(CommonDataKeys.VIRTUAL_FILE) ?: return
-        val module = ModuleUtilCore.findModuleForFile(file, project)
-        if (module == null) {
-            ActionUtil.isEnabledAndVisible(e, false)
-            return
-        }
-        if (file.canonicalPath != project.basePath || file != project.guessProjectDir()) {
-            ActionUtil.isEnabledAndVisible(e, false)
-            return
-        }
+        val module = ModuleUtilCore.findModuleForFile(file, project) ?: return
         val activeBeans = SpringSearchServiceFacade.getInstance(project).getAllActiveBeans(module)
-        p.isEnabledAndVisible = activeBeans.isNotEmpty()
+        p.setVisible(activeBeans.isNotEmpty())
     }
 
     override fun getActionUpdateThread(): ActionUpdateThread {
