+++ modules/spring-core/src/main/kotlin/com/explyt/spring/core/completion/properties/ProjectConfigurationPropertiesLoader.kt
@@ -28,7 +28,6 @@ import com.explyt.util.ExplytPsiUtil.returnPsiClass
 import com.explyt.util.runReadNonBlocking
 import com.intellij.json.psi.JsonFile
 import com.intellij.openapi.module.Module
-import com.intellij.openapi.module.ModuleUtilCore
 import com.intellij.openapi.project.Project
 import com.intellij.psi.*
 import com.intellij.psi.javadoc.PsiDocToken
@@ -98,15 +97,15 @@ class ProjectConfigurationPropertiesLoader(project: Project) :
     ): Map<String, ConfigurationProperty> {
         return CachedValuesManager.getManager(module.project).getCachedValue(annotatedElement) {
             CachedValueProvider.Result(
-                loadPropertiesFromConfigurationFile(annotatedElement), annotatedElement
+                loadPropertiesFromConfigurationFile(module, annotatedElement), annotatedElement
             )
         }
     }
 
     private fun loadPropertiesFromConfigurationFile(
-        annotatedElement: PsiModifierListOwner
+        module: Module,
+        annotatedElement: PsiModifierListOwner,
     ): Map<String, ConfigurationProperty> {
-        val module = ModuleUtilCore.findModuleForPsiElement(annotatedElement) ?: return emptyMap()
         val result = hashMapOf<String, ConfigurationProperty>()
         val prefix = ProjectConfigurationPropertiesUtil
             .extractConfigurationPropertyPrefix(module, annotatedElement) ?: return emptyMap()
