+++ modules/spring-core/src/main/kotlin/com/explyt/spring/core/properties/contributors/UastConfigurationPropertyReferenceContributor.kt
@@ -63,6 +63,6 @@ class UastConfigurationPropertyReferenceContributor : PsiReferenceContributor() 
     }
 
     companion object {
-        val VALUE_SCHEDULED = listOf("cron", "initialDelayString", "fixedDelayString", "fixedRateString", "zone")
+        val VALUE_SCHEDULED = listOf("cron", "initialDelayString", "fixedDelayString", "fixedRateString")
     }
 }
