+++ modules/spring-core/src/main/kotlin/com/explyt/spring/core/providers/SpringBeanLineMarkerProviderNativeLibrary.kt
@@ -25,7 +25,6 @@ import com.explyt.spring.core.providers.SpringBeanLineMarkerProvider.Companion.i
 import com.explyt.spring.core.providers.SpringBeanLineMarkerProvider.Companion.isAutowiredMethodExpression
 import com.explyt.spring.core.service.NativeSearchService
 import com.explyt.spring.core.service.PsiBean
-import com.explyt.spring.core.service.SpringSearchService
 import com.explyt.spring.core.service.SpringSearchUtils
 import com.explyt.spring.core.statistic.StatisticActionId
 import com.explyt.spring.core.statistic.StatisticService
@@ -44,7 +43,6 @@ import com.intellij.openapi.editor.markup.GutterIconRenderer
 import com.intellij.openapi.util.NotNullLazyValue
 import com.intellij.psi.*
 import org.jetbrains.uast.*
-import java.util.function.Supplier
 
 class SpringBeanLineMarkerProviderNativeLibrary : RelatedItemLineMarkerProvider() {
 
@@ -53,22 +51,13 @@ class SpringBeanLineMarkerProviderNativeLibrary : RelatedItemLineMarkerProvider(
         result: MutableCollection<in RelatedItemLineMarkerInfo<*>>
     ) {
         val uClass = getUParentForIdentifier(element) as? UClass ?: return
+
         val psiClass = uClass.javaPsi
         if (!isSpringBeanCandidateClass(psiClass)) return
-        val isComponentCandidate = isComponentCandidate(psiClass)
-        val project = element.project
-        val libraryBeans = NativeSearchService.getInstance(project).getAllProjectNodesLibraryBeans()
-        if (libraryBeans.isEmpty()) {
-            val beanSupplier = { SpringSearchService.getInstance(project).getAllActiveBeans().toList() }
-            if (isComponentCandidate) {
-                addContextBean(uClass, false, result, beanSupplier)
-            }
-            processMethods(uClass, result, beanSupplier)
-            processFields(uClass, result, beanSupplier)
-            return
-        }
+        val libraryBeans = NativeSearchService.getInstance(element.project).getAllProjectNodesLibraryBeans()
         val targetQualifiedName = psiClass.qualifiedName
         val contextBean = libraryBeans.find { isContextClass(it, targetQualifiedName) }
+        val isComponentCandidate = isComponentCandidate(psiClass)
 
         if (contextBean?.psiMember is PsiMethod) {
             addMethodBeanDeclaration(uClass, contextBean, result)
@@ -76,9 +65,9 @@ class SpringBeanLineMarkerProviderNativeLibrary : RelatedItemLineMarkerProvider(
 
         if (contextBean == null && !isComponentCandidate) return
 
-        addContextBean(uClass, contextBean == null, result) { libraryBeans }
-        processMethodsNative(uClass, result) { libraryBeans }
-        processFields(uClass, result) { libraryBeans }
+        addContextBean(uClass, libraryBeans, contextBean == null && isComponentCandidate, result)
+        processMethods(uClass, libraryBeans, result)
+        processFields(uClass, libraryBeans, result)
     }
 
     private fun isContextClass(it: PsiBean, targetQualifiedName: String?): Boolean {
@@ -91,9 +80,10 @@ class SpringBeanLineMarkerProviderNativeLibrary : RelatedItemLineMarkerProvider(
 
     private fun processFields(
         uClass: UClass,
-        result: MutableCollection<in RelatedItemLineMarkerInfo<*>>,
-        supplier: Supplier<List<PsiBean>>
+        libraryBeans: List<PsiBean>,
+        result: MutableCollection<in RelatedItemLineMarkerInfo<*>>
     ) {
+
         val fields = uClass.fields.takeIf { it.isNotEmpty() } ?: return
         for (uField in fields) {
             val psiField = uField.javaPsi as? PsiField ?: continue
@@ -102,7 +92,7 @@ class SpringBeanLineMarkerProviderNativeLibrary : RelatedItemLineMarkerProvider(
             val sourcePsi = uField.uastAnchor?.sourcePsi ?: continue
             val builder = NavigationGutterIconBuilder.create(SpringIcons.SpringBeanDependencies)
                 .setAlignment(GutterIconRenderer.Alignment.LEFT)
-                .setTargets(NotNullLazyValue.lazy { getBeanDeclarations(uField, supplier.get()) })
+                .setTargets(NotNullLazyValue.lazy { getBeanDeclarations(uField, libraryBeans) })
                 .setTooltipText(SpringCoreBundle.message("explyt.spring.gutter.tooltip.title.choose.bean.candidate"))
                 .setPopupTitle(SpringCoreBundle.message("explyt.spring.gutter.popup.title.choose.bean.candidate"))
                 .setEmptyPopupText(SpringCoreBundle.message("explyt.spring.gutter.notfound.title.choose.bean.candidate"))
@@ -110,19 +100,18 @@ class SpringBeanLineMarkerProviderNativeLibrary : RelatedItemLineMarkerProvider(
         }
     }
 
-    private fun processMethodsNative(
+    private fun processMethods(
         uClass: UClass,
-        result: MutableCollection<in RelatedItemLineMarkerInfo<*>>,
-        beanSupplier: Supplier<List<PsiBean>>
+        libraryBeans: List<PsiBean>,
+        result: MutableCollection<in RelatedItemLineMarkerInfo<*>>
     ) {
-        val libraryBeans = beanSupplier.get()
         val beanMethodNames = libraryBeans.asSequence()
             .filter { it.psiMember is PsiMethod && it.psiMember.containingClass?.qualifiedName == uClass.qualifiedName }
             .mapToSet { (it.psiMember as PsiMethod).name }
         for (method in uClass.methods) {
             val psiElement = method.uastAnchor?.sourcePsi ?: continue
             if (method.isConstructor || isAutowiredMethodExpression(method.javaPsi)) {
-                checkMethodParameters(method, result, beanSupplier)
+                checkMethodParameters(method, libraryBeans, result)
                 if (beanMethodNames.isEmpty()) return
                 continue
             }
@@ -137,45 +126,15 @@ class SpringBeanLineMarkerProviderNativeLibrary : RelatedItemLineMarkerProvider(
                     .setEmptyPopupText(SpringCoreBundle.message("explyt.spring.gutter.notfound.title.choose.autowired.candidate"))
                     .setTargetRenderer { SpringBeanLineMarkerProvider().getTargetRender() }
                 result.add(builder.createLineMarkerInfo(psiElement))
-                checkMethodParameters(method, result, beanSupplier)
+                checkMethodParameters(method, libraryBeans, result)
             }
         }
     }
-
-    private fun processMethods(
-        uClass: UClass,
-        result: MutableCollection<in RelatedItemLineMarkerInfo<*>>,
-        beanSupplier: Supplier<List<PsiBean>>
-    ) {
-
-        for (method in uClass.methods) {
-            val psiElement = method.uastAnchor?.sourcePsi ?: continue
-            if (method.isConstructor || isAutowiredMethodExpression(method.javaPsi)) {
-                checkMethodParameters(method, result, beanSupplier)
-                continue
-            }
-
-            if (!method.javaPsi.isMetaAnnotatedBy(SpringCoreClasses.BEAN)) continue
-
-            val builder = NavigationGutterIconBuilder.create(SpringIcons.SpringBean)
-                .setAlignment(GutterIconRenderer.Alignment.LEFT)
-                .setTargets(NotNullLazyValue.lazy {
-                    findFieldsAndMethodsWithAutowired(null, method, beanSupplier.get())
-                })
-                .setTooltipText(SpringCoreBundle.message("explyt.spring.gutter.tooltip.title.choose.autowired.candidate"))
-                .setPopupTitle(SpringCoreBundle.message("explyt.spring.gutter.popup.title.choose.autowired.candidate"))
-                .setEmptyPopupText(SpringCoreBundle.message("explyt.spring.gutter.notfound.title.choose.autowired.candidate"))
-                .setTargetRenderer { SpringBeanLineMarkerProvider().getTargetRender() }
-            result.add(builder.createLineMarkerInfo(psiElement))
-            checkMethodParameters(method, result, beanSupplier)
-        }
-    }
-
 
     private fun checkMethodParameters(
         method: UMethod,
-        result: MutableCollection<in RelatedItemLineMarkerInfo<*>>,
-        beanSupplier: Supplier<List<PsiBean>>
+        libraryBeans: List<PsiBean>,
+        result: MutableCollection<in RelatedItemLineMarkerInfo<*>>
     ) {
         val uastParameters = method.uastParameters.takeIf { it.isNotEmpty() } ?: return
 
@@ -183,7 +142,7 @@ class SpringBeanLineMarkerProviderNativeLibrary : RelatedItemLineMarkerProvider(
             val sourcePsi = uParameter.uastAnchor?.sourcePsi ?: continue
             val builder = NavigationGutterIconBuilder.create(SpringIcons.SpringBeanDependencies)
                 .setAlignment(GutterIconRenderer.Alignment.LEFT)
-                .setTargets(NotNullLazyValue.lazy { getBeanDeclarations(uParameter, beanSupplier.get()) })
+                .setTargets(NotNullLazyValue.lazy { getBeanDeclarations(uParameter, libraryBeans) })
                 .setTooltipText(SpringCoreBundle.message("explyt.spring.gutter.tooltip.title.choose.bean.candidate"))
                 .setPopupTitle(SpringCoreBundle.message("explyt.spring.gutter.popup.title.choose.bean.candidate"))
                 .setEmptyPopupText(SpringCoreBundle.message("explyt.spring.gutter.notfound.title.choose.bean.candidate"))
@@ -193,9 +152,9 @@ class SpringBeanLineMarkerProviderNativeLibrary : RelatedItemLineMarkerProvider(
 
     private fun addContextBean(
         uClass: UClass,
+        libraryBeans: List<PsiBean>,
         isComponentCandidate: Boolean,
-        result: MutableCollection<in RelatedItemLineMarkerInfo<*>>,
-        beanSupplier: Supplier<List<PsiBean>>
+        result: MutableCollection<in RelatedItemLineMarkerInfo<*>>
     ) {
         val sourcePsi = uClass.uastAnchor?.sourcePsi ?: return
         val icon = if (isComponentCandidate) SpringIcons.springBeanInactive else SpringIcons.SpringBean
@@ -204,7 +163,7 @@ class SpringBeanLineMarkerProviderNativeLibrary : RelatedItemLineMarkerProvider(
             SpringCoreBundle.message("explyt.spring.gutter.tooltip.title.choose.autowired.candidate")
         val builder = NavigationGutterIconBuilder.create(icon)
             .setAlignment(GutterIconRenderer.Alignment.LEFT)
-            .setTargets(NotNullLazyValue.lazy { findFieldsAndMethodsWithAutowired(uClass, null, beanSupplier.get()) })
+            .setTargets(NotNullLazyValue.lazy { findFieldsAndMethodsWithAutowired(uClass, null, libraryBeans) })
             .setTooltipText(text)
             .setPopupTitle(SpringCoreBundle.message("explyt.spring.gutter.popup.title.choose.autowired.candidate"))
             .setEmptyPopupText(SpringCoreBundle.message("explyt.spring.gutter.notfound.title"))
@@ -244,9 +203,7 @@ class SpringBeanLineMarkerProviderNativeLibrary : RelatedItemLineMarkerProvider(
             JavaEeClasses.INJECT.allFqns + JavaEeClasses.RESOURCE.allFqns + SpringCoreClasses.AUTOWIRED
 
         val nativeSearchService = NativeSearchService.getInstance(project)
-        val projectBeans = nativeSearchService.getProjectBeans()
-        val allBeans = if (projectBeans.isNotEmpty()) libraryBeans + projectBeans
-        else SpringSearchService.getInstance(project).getAllActiveBeans()
+        val allBeans = libraryBeans + nativeSearchService.getProjectBeans()
 
         val allFieldsWithAutowired = allBeans.asSequence()
             .mapNotNull { bean -> bean.psiClass.toUElementOfType<UClass>()?.fields }
