+++ modules/spring-core/src/main/resources/messages/SpringCoreBundle.properties
@@ -216,5 +216,3 @@ explyt.spring.gutter.spi.tooltip.title.usage=Navigate to SPI
 explyt.spring.gutter.spi.popup.title.usage=Choose SPI implementation
 explyt.spring.folder.mark.property=Spring Application Property Folder
 explyt.spring.folder.unmark.property=Unmark as Spring Application Property Folder
-explyt.spring.debugger.show.dialog=\u2026Evaluate
-explyt.spring.debugger.root.node=You can use: 'Explyt.context' or 'Explyt.getBeanFactory' or 'Explyt.getEnvironment' in Evaluation Expression to access the Spring Context
