+++ modules/spring-core/src/test/kotlin/com/explyt/spring/core/inspections/java/SpringPropertiesInspectionTest.kt
@@ -20,7 +20,6 @@ package com.explyt.spring.core.inspections.java
 import com.explyt.spring.core.inspections.SpringPropertiesInspection
 import com.explyt.spring.test.ExplytInspectionJavaTestCase
 import com.explyt.spring.test.TestLibrary
-import org.intellij.lang.annotations.Language
 import org.jetbrains.kotlin.test.TestMetadata
 
 class SpringPropertiesInspectionTest : ExplytInspectionJavaTestCase() {
@@ -49,34 +48,12 @@ class SpringPropertiesInspectionTest : ExplytInspectionJavaTestCase() {
         myFixture.testHighlighting("application.properties")
     }
 
-    fun testKebabCaseProperties() {
+    fun testNoDuplicateProperties() {
         myFixture.configureByText(
             "application.properties",
             """
-            <warning descr="Should be kebab-case">test.fooBar</warning>=1
-            test.foo-bar1=2
-            """.trimIndent()
-        )
-        myFixture.testHighlighting("application.properties")
-    }
-
-    fun testFileDefinitionProperties() {
-        @Language("java") val configurationProperty = """
-            @org.springframework.context.annotation.Configuration
-            @org.springframework.boot.context.properties.ConfigurationProperties(prefix = "explyt.prop")
-            public class ConfigProperties {               
-                private String test;                           
-                public void setTest(String test) { this.test = test}
-                public String getTest() { return this.test}
-            } 
-        """.trimIndent()
-        myFixture.addClass(configurationProperty)
-        myFixture.configureByText(
-            "application.properties",
-            """
-            explyt.prop.test=test                
-            <warning descr="Cannot resolve key property 'explyt.prop.field'">explyt.prop.field</warning>=test            
-            test.foo-bar1=2
+            <warning descr="Cannot resolve key property 'test.fooBar'"><warning descr="Should be kebab-case">test.fooBar</warning></warning>=1
+            <warning descr="Cannot resolve key property 'test.foo-bar1'">test.foo-bar1</warning>=2
             """.trimIndent()
         )
         myFixture.testHighlighting("application.properties")
