+++ modules/spring-web/src/main/kotlin/com/explyt/spring/web/action/SpringWebJavaMethodGenerateAction.kt
@@ -17,7 +17,7 @@
 
 package com.explyt.spring.web.action
 
-import com.explyt.spring.core.action.JavaCoreMethodGenerateUtils
+import com.explyt.spring.core.action.JavaMethodGenerateUtils
 import com.explyt.spring.core.statistic.StatisticActionId
 import com.explyt.spring.core.statistic.StatisticService
 import com.explyt.spring.web.SpringWebBundle
@@ -31,7 +31,6 @@ import com.explyt.spring.web.parser.HttpMethod
 import com.explyt.spring.web.parser.HttpParamType
 import com.explyt.spring.web.parser.UrlParser
 import com.explyt.spring.web.util.SpringWebUtil
-import com.explyt.util.JavaMethodGenerateUtils
 import com.intellij.codeInsight.CodeInsightActionHandler
 import com.intellij.codeInsight.generation.actions.BaseGenerateAction
 import com.intellij.codeInsight.template.Template
@@ -95,7 +94,7 @@ private class WebMethodHandler : CodeInsightActionHandler {
                 CurlParser.parse(urlString) else UrlParser.parse(urlString)
 
             runWriteAction {
-                val targetClass = JavaCoreMethodGenerateUtils.getNearTargetClass(editor, file) ?: return@runWriteAction
+                val targetClass = JavaMethodGenerateUtils.getNearTargetClass(editor, file) ?: return@runWriteAction
                 val documentManager = PsiDocumentManager.getInstance(project)
                 val document = documentManager.getDocument(file) ?: return@runWriteAction
                 PsiDocumentManager.getInstance(file.project).commitDocument(document)
