+++ modules/spring-web/src/main/kotlin/com/explyt/spring/web/action/SpringWebKotlinMethodGenerateAction.kt
@@ -17,6 +17,7 @@
 
 package com.explyt.spring.web.action
 
+import com.explyt.spring.core.action.KotlinMethodGenerateUtils
 import com.explyt.spring.core.statistic.StatisticActionId
 import com.explyt.spring.core.statistic.StatisticService
 import com.explyt.spring.web.SpringWebBundle
@@ -30,7 +31,6 @@ import com.explyt.spring.web.parser.HttpMethod
 import com.explyt.spring.web.parser.HttpParamType
 import com.explyt.spring.web.parser.UrlParser
 import com.explyt.spring.web.util.SpringWebUtil
-import com.explyt.util.KotlinMethodGenerateUtils
 import com.intellij.codeInsight.CodeInsightActionHandler
 import com.intellij.codeInsight.generation.actions.BaseGenerateAction
 import com.intellij.codeInsight.template.Template
